+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/PlatformRequester.cs
@@ -326,10 +326,14 @@ namespace Beamable.Api
 				request.SetRequestHeader(Constants.Requester.HEADER_PID, Pid);
 			}
 
-			request.SetRequestHeader(Constants.Requester.HEADER_BEAMABLE_VERSION, "");
+			#if !BEAMABLE_DISABLE_VERSION_HEADERS
+			request.SetRequestHeader(Constants.Requester.HEADER_BEAMABLE_VERSION, _beamableVersion.ToString());
 			request.SetRequestHeader(Constants.Requester.HEADER_APPLICATION_VERSION, Application.version);
 			request.SetRequestHeader(Constants.Requester.HEADER_UNITY_VERSION, Application.unityVersion);
-
+			request.SetRequestHeader(Constants.Requester.HEADER_ENGINE_TYPE, Application.isEditor
+				                         ? Constants.Requester.USER_AGENT_UNITY_EDITOR
+				                         : Constants.Requester.USER_AGENT_UNITY);
+			#endif
 
 			if (includeAuthHeader)
 			{